+++ bindings/js/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to EnhEx JavaScript binding will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [js-v0.4.0] - 2026-06-01
+
+- Add tagged template support with ``enhex`...` ``
+
 ## [js-v0.3.1] - 2026-05-29
 
 - Add missing npm package
@@ -17,5 +21,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - TypeScript types
 - Test suite
 
+[js-v0.4.0]: https://github.com/mkh-user/enhex/releases/tag/js-v0.4.0
 [js-v0.3.1]: https://github.com/mkh-user/enhex/releases/tag/js-v0.3.1
 [js-v0.3.0]: https://github.com/mkh-user/enhex/releases/tag/v0.3.0